size_t -> unsigned

+++ dtrain/pairsampling.h
@@ -11,8 +11,8 @@ namespace dtrain
 inline void
 sample_all_pairs(vector<ScoredHyp>* s, vector<pair<ScoredHyp,ScoredHyp> > &training)
 {
-  for (size_t i = 0; i < s->size()-1; i++) {
-    for (size_t j = i+1; j < s->size(); j++) {
+  for (unsigned i = 0; i < s->size()-1; i++) {
+    for (unsigned j = i+1; j < s->size(); j++) {
       pair<ScoredHyp,ScoredHyp> p;
       p.first = (*s)[i];
       p.second = (*s)[j];
@@ -25,8 +25,8 @@ inline void
 sample_rand_pairs(vector<ScoredHyp>* s, vector<pair<ScoredHyp,ScoredHyp> > &training,
                   MT19937* prng)
 {
-  for (size_t i = 0; i < s->size()-1; i++) {
-    for (size_t j = i+1; j < s->size(); j++) {
+  for (unsigned i = 0; i < s->size()-1; i++) {
+    for (unsigned j = i+1; j < s->size(); j++) {
       if (prng->next() < .5) {
         pair<ScoredHyp,ScoredHyp> p;
         p.first = (*s)[i];
